Surfaces: improve size detection when wrapping DCs
For DCs associated with an object, you now have three choices for
determining size:

1) Pass the hWnd associated with the DC.  This is preferred, as PD can
auto-magically pick up size changes with no work on the callers part.
2) Manually specify width/height.  This is really the only option for
DCs not associated with a window (e.g. printer DCs)
3) Don't pass any sizes.  This is okay as long as you don't need to
query the surface's size for anything (e.g. if you're just drawing items
with hard-coded size onto the surface).  If you do need the surface's
size, this method is *not* recommended, as DCs don't have dimensions!
As a fallback, PD will try to query the currently selected bitmap, but
this is not guaranteed to work.

Note that none of this crap is required for internally managed surfaces
- only when wrapping a DC for temporary painting.