Start work on documentation
For now, my main focus is a short "Getting Started" guide that walks you
through the basic elements of the library.

As an FYI, comprehensive documentation is unlikely to happen (given the
size of the code base), but most functions are designed to be
self-explanatory, so I'm not too worried about this.