Large cross-update from PhotoDemon
pd2D has received a lot of stress-testing over the past year, both from
PhotoDemon itself as well as a number of client projects.  While I'm
still not satisfied with certain elements of its design, it can
undoubtedly save a ton of time when implementing graphics tasks in VB,
and I feel comfortable calling it "stable enough for general use".

I've made a large number of changes, improvements, and bug-fixes to the
library over the past year, but I haven't been good about syncing these
changes back to this repository.  As of this commit, however, this
standalone version is *finally* back to parity with the version shipped
in PhotoDemon.