Copy over latest PhotoDemon versions of pd2D
Will I ever attempt to document this library?  Probably not.  But it bothers me that everything here is so out-of-date.

So this is a quick-and-dirty dump of all pd2D drawing modules in PhotoDemon.  The library design is much-improved over old versions, with better performance and cleaner APIs.  Note that there is potentially a *ton* of unused code in this commit because all these changes were blindly copy+pasted from PD.  I'll look at removing unnecessary code in the coming days.